The game works!!! Yes! A minor problem is specified below, but other than that the game works as intended!
Intermediate Problem:
- [] Unit testing is change:  Before in the logical grid the first player pawns where in the top row and the second player pawns where in the bottom row. This did not make sense if you think from the perspective of the viewer/player, now the first players pawns start at the bottom row and the second player pawns start at the top row. With this change the test will be different.
- [] Ensure that all the comments match the code:  We both need to double check this since it is the most common way I get points marked off.

Minor Problem:
- [] When you click outside the board the board disappears. Probably has something to do with controller.cpp, but we should ask about this in class because the fix is not obvious.

Very Minor Problem:
- [] Get Fancy?: Add a game over text when the game is over. I tried to do this really quick, but it crashed the function... I need to work on other projects/hw so you can add this if you want @Quinciahu.

------------------------------------  Previous Problems are essentially fixed ------------------------------------------
Major Problem (but fixable):
FIXED: If a player gets to the other side. The game quits w/ exit code 11
FIXED: If the first piece to move is not white. The game quits w/ exit code 11

Intermediate Problem:
FIXED: If you select a piece and you want to select another piece, instead of putting the piece back to previous position to deselect, you must first select the piece you wish to switch to and this will deselect... clunky.
Intended: If it is a stalemate the game is frozen in frame

Minor Problem:
FIXED: The colors for ge211 are way off so we have to pick 4 colors that contrast so that the board/pieces will not look like shit,
Ask about during class: There is also a problem with the draw function in view.h, I had to code it especially clunky because of some mysterious errors that ge211 gave off. Before submit must reduce the if/else conditions because Rob/Jess will definitely comment on it.